requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1

# ── Data generation ───────────────────────────────────────────────────────────
faker==25.2.0
numpy==1.26.4

# ── Kafka ─────────────────────────────────────────────────────────────────────
kafka-python==2.0.2

# ── Storage & Query ───────────────────────────────────────────────────────────
duckdb==0.10.3
pandas==2.2.2
pyarrow==16.1.0
fastparquet==2024.2.0

# ── Anomaly Detection ─────────────────────────────────────────────────────────
scikit-learn==1.5.0
scipy==1.13.1

# ── Forecasting ───────────────────────────────────────────────────────────────
statsmodels==0.14.2
pmdarima==2.0.4          # auto-ARIMA order selection

# ── MLflow ────────────────────────────────────────────────────────────────────
mlflow==2.13.2

# ── Airflow ───────────────────────────────────────────────────────────────────
# Install via Astro CLI — not pip directly
# apache-airflow==2.9.1

# ── Alerting ──────────────────────────────────────────────────────────────────
slack-sdk==3.27.2
requests==2.32.3

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0

# ── Linting ───────────────────────────────────────────────────────────────────
ruff==0.4.8
black==24.4.2